+++ CHANGES.md
@@ -3,8 +3,9 @@
 * BREAKING -- `updatePlaceholder` is now `updateProgress`, with same arguments
 * BREAKING -- `{status, progress, failed}` will not be set from initial content or `setContent`: you must call `updateProgress`
 * BREAKING -- `ed.updateProgress` updates internally and does not change content block metadata (#237)
-* `ed.updateProgress(id, {progress: null})` will now remove progress bar
 * Simplifying updates = perf improvements with `updateProgress`
+* `ed.updateProgress(id, {progress: null})` will now remove progress bar
+* `ed.updateProgress(id, {failed: true})` will now show an error message and retry button
 
 ## 0.17.8 - 2016-07-14